% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkpData.R
\name{checkpData}
\alias{checkpData}
\title{Checks the sampleinfo file
Reads the excelsheet containing sample information, performs sanity checks and generates pData}
\usage{
checkpData(file, type)
}
\arguments{
\item{file}{: path to the inputfile}

\item{type}{: Input type. Can be a dataframe or a csv file. Options are df and csv}
}
\value{
Dataframe with pData
}
\description{
Checks the sampleinfo file
Reads the excelsheet containing sample information, performs sanity checks and generates pData
}
\examples{
checkpData(file="/home/user/file.csv",type="csv")
checkpData(file=dataframe,type="df")
}
