% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cleanup.R
\name{Cleanup}
\alias{Cleanup}
\title{Cleans up topTable results
converts logFC to FC}
\usage{
Cleanup(tt)
}
\arguments{
\item{vals}{: output from topTable}
}
\value{
Dataframe with limma output
}
\description{
Cleans up topTable results
converts logFC to FC
}
\examples{
Cleanup()
}
