% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cleanup2.R
\name{Cleanup2}
\alias{Cleanup2}
\title{Cleans up topTable results
converts logFC to FC}
\usage{
Cleanup2(tt)
}
\arguments{
\item{vals}{: output from topTable}
}
\value{
Dataframe with limma output
}
\description{
Cleans up topTable results
converts logFC to FC
}
\examples{
Cleanup2()
}
