% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateCameraData.R
\name{CreateCameraData}
\alias{CreateCameraData}
\title{Runs camera}
\usage{
CreateCameraData(limma, voom, dataset, design)
}
\arguments{
\item{voom}{: eset having expression data}

\item{dataset}{: Molecular signature database gene set ("Mm" for mouse and "Hs" for humans). Currently only Hallmark and Curated exist}

\item{design}{: model matrix}

\item{Limma}{: Limma data}
}
\value{
Camera data and indices
}
\description{
Runs camera
}
\examples{
CreateCameraData(limma,voom,dataset,design)
}

