% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PerGroupAvg.R
\name{PerGroupAvg}
\alias{PerGroupAvg}
\title{Computes per group average expression}
\usage{
PerGroupAvg(eset)
}
\arguments{
\item{eset}{: eset}
}
\value{
Dataframe with per group average expression
}
\description{
Computes per group average expression
}
\examples{
PerGroupAvg()
}
