% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runTopGO.R
\name{runTopGO}
\alias{runTopGO}
\title{Return a data frame of gene annotations
Takes a dataframe having output from limma, cleans up to remove unwanted columns and annotates it}
\usage{
runTopGO(tt, pcutoff = 0.1)
}
\arguments{
\item{ids}{: List of ENSEMBL IDs}

\item{organism}{: Mm for Mus musculus (mouse) and Hs for Homo sapiens (humans)}
}
\value{
Dataframe with limma data and the annotation
}
\description{
Return a data frame of gene annotations
Takes a dataframe having output from limma, cleans up to remove unwanted columns and annotates it
}

