% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svaBatchCor.R
\name{svaBatchCor}
\alias{svaBatchCor}
\title{Runs sva batch correction
converts logFC to FC}
\usage{
svaBatchCor(dat, mmi, mm0, n.sv = NULL)
}
\arguments{
\item{dat}{: matrix of cpms}

\item{mmi}{: model matrix for fitting the data}

\item{mm0}{: null model being compared when fitting the data}

\item{n.sv}{: number of significant surrogate variables}
}
\value{
Matrix of batch corrected cpms
}
\description{
Runs sva batch correction
converts logFC to FC
}
\examples{
mod <- model.matrix(~as.factor(Disease), data=pData)
mod0 <- model.matrix(~1,data=pData)
n.sv = num.sv (data, mod,method = "leek")
cpms.corr <- svaBatchCor(cpms,mod,mod0,n.sv=n.sv)$corrected
}
