% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneAnnotate.R
\name{GeneAnnotate}
\alias{GeneAnnotate}
\title{Return a data frame of gene annotations
Takes a dataframe having output from limma, cleans up to remove unwanted columns and annotates it}
\usage{
GeneAnnotate(ids, organism)
}
\arguments{
\item{ids}{: List of ENSEMBL IDs}

\item{organism}{: mouse for Mus musculus (mouse) Rat for Rattus norvegicus (Rat) and human for Homo sapiens (humans)}
}
\value{
Dataframe with limma data and the annotation
}
\description{
Return a data frame of gene annotations
Takes a dataframe having output from limma, cleans up to remove unwanted columns and annotates it
}
