% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runTopGO.R
\name{runTopGO}
\alias{runTopGO}
\title{Return a data frame of gene annotations
Takes a dataframe having output from limma, cleans up to remove unwanted columns and annotates it}
\usage{
runTopGO(tt, organism, pcutoff = 0.1)
}
\arguments{
\item{organism}{: mouse for Mus musculus (mouse) Rat for Rattus norvegicus (Rat) and human for Homo sapiens (humans)}

\item{ids}{: List of ENSEMBL IDs}
}
\value{
Dataframe with limma data and the annotation
}
\description{
Return a data frame of gene annotations
Takes a dataframe having output from limma, cleans up to remove unwanted columns and annotates it
}
